backend/requirements.txt
# Flask and extensions
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
Flask-SocketIO==5.3.6
flask-jwt-extended==4.6.0

# Async server
eventlet==0.34.2
dnspython<2.7.0

# Database
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1

# Redis
redis==5.0.1

# Security
bcrypt==4.1.2
cryptography==41.0.7
PyJWT==2.8.0
pyotp==2.9.0
qrcode==7.4.2

# Validation
pydantic==2.5.3
email-validator==2.1.0

# Storage
boto3==1.34.11

# PDF Generation
WeasyPrint==60.2
pydyf==0.10.0

# AI - use compatible versions
openai==1.12.0
google-generativeai==0.4.0

# Supabase - use newer version with relaxed httpx requirement
supabase==2.4.0

# HTTP requests - let dependencies resolve this
httpx>=0.24.0,<0.28.0
requests==2.31.0

# Utilities
python-dotenv==1.0.0
python-dateutil==2.8.2
Pillow==10.1.0

# Testing
pytest==7.4.4
pytest-cov==4.1.0
pytest-asyncio==0.23.3

# Type hints
typing-extensions==4.9.0

# Data processing
pandas==2.2.0
openpyxl==3.1.2

# Security
Flask-Limiter
flask-talisman
